+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## [0.1.43]
+* Update GetServicePrincipalToken to allow using certificate for App Reg authentication
 * Update deps
 
 ## [0.1.42]